Add detailed view for bits and force history tracking
Implemented a new "Detail View" for specific bits, displaying metadata, comments, force information, and a history of the last five force actions. Introduced "View Details" buttons in resource tables and modified underlying logic to support dynamic data fetching. Enhanced UI with consistent styling and responsive tables for better usability.
